services/api-gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
httpx==0.27.2
redis==5.0.7
pydantic-settings==2.5.2
python-multipart==0.0.9
orjson==3.10.12
prometheus-fastapi-instrumentator==6.0.0

# Authentication
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Email validation
email-validator==2.1.0

# Additional utilities
python-dateutil==2.8.2

